Move DockerService functions to OutputParser.
